Make the capability settings shell the only admin UI
Remove the dual-shell machinery: the is_shell_v2_active() gate, the
gtmkit_shell_v2 filter and the ?gtmkit_shell request override. The
settings menu now registers a single top-level entry instead of a
per-page submenu, and the legacy option-page classes that only powered
that submenu (Integrations, Help, Templates, Upgrades) are deleted.

Relocate get_plugins() to the Integrations helper so the shell bootstrap
keeps its integration data, and add a gtmkit_settings filter so add-ons
can contribute to the localized settings payload (for example license
details) instead of rewriting the enqueued script.
