+++ changelog.txt
@@ -1,8 +1,9 @@
 *** GTM Kit ***
 
 Unreleased
-* Changed: The settings screen now uses the redesigned, capability-based interface by default, with settings grouped into Setup, Events & data layer, Commerce, Consent & privacy, Tools and more. The previous layout stays available as a fallback: append `?gtmkit_shell=v1` to a settings URL for a single visit, or use the `gtmkit_shell_v2` filter to switch a whole site back.
+* Changed: The settings screen now uses the redesigned, capability-based interface, with settings grouped into Setup, Events & data layer, Commerce, Consent & privacy, Tools and more. The previous per-page settings layout has been removed; the redesigned interface is the only settings UI.
 * Fix: The Contact Form 7 "Load JavaScript" setting now shows the recommended choice as selected when the setting has never been saved.
+* Dev: New `gtmkit_settings` filter lets add-ons add data to the settings screen's localized payload (such as license details) without rewriting the enqueued script.
 * Dev: New `gtmkit_option_value` filter runs on every option read, including unknown keys, so add-ons can resolve a value from an alternative source (for example a network-level override) without changing the core read path.
 
 2026-06-12 - version 2.15.0
